requirements.txt
channels==4.0.0
daphne==4.0.0
Django==4.2.7
django-cors-headers==4.3.0
django-debug-toolbar==4.2.0
django-phonenumber-field==7.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
flake8==3.9.2
phonenumbers==8.13.24
psycopg2-binary==2.9.9
PyJWT==2.8.0
pyOpenSSL==24.0.0
requests==2.31.0
sentry-sdk==1.1.0
twilio==8.13.0